第一支API -  傳送header & 接收resList參數(使用gson)

+++ springboot-javafx/src/main/java/com/demo/model/MnrResponse.java
@@ -4,7 +4,7 @@ package com.demo.model;
 public class MnrResponse {
 
     // 回應參數
-    private int systemUpdateTime;
+    private long systemUpdateTime;
     private int port;
     private String passportNo;
     private String chineseName;
@@ -12,7 +12,7 @@ public class MnrResponse {
     private String applyDate;
 
 
-    public int getSystemUpdateTime() {
+    public long getSystemUpdateTime() {
         return systemUpdateTime;
     }
